Recorder packet data and its size when its being sent
One of the main reasons to hook into the client is that we can figure out which
bytes of the packet were submitted together. This allows us to easily find the
size of a given element. With this, we now write down this information by
grouping the sent bytes according to the size the client puts in.