Also include received packets
We currently only log packets the client is trying to send, because they were
easier to collect. This now extends the mechanism to also collect packets being
received. Technically, this should help figuring out the structure for a packet
the client is expecting. However, if the packet does not have the correct
structure, the client will most likely crash before we can note down what the
client wanted. To ensure that usecase is covered, we'll have to send off the log
 to the injecting programing, as that would (hopefully) stay running.